requirements.txt
Django==5.2.5
gql==3.5.3
PyGithub==2.6.1
PyGithub==2.7.0
python-dotenv==1.1.1
